Revert "[ETCM-893] Add support for chains with ids bigger than 127 (#1006)"
This reverts commit 3982b037b9c6a4c5ca7562585626d6fd6550127a.

+++ src/it/scala/io/iohk/ethereum/txExecTest/ECIP1017Test.scala
@@ -19,7 +19,7 @@ class ECIP1017Test extends AnyFlatSpec with Matchers {
       monetaryPolicyConfig = MonetaryPolicyConfig(EraDuration, 0.2, 5000000000000000000L, 3000000000000000000L),
       // unused
       maxCodeSize = None,
-      chainId = 0x3d,
+      chainId = 0x3d.toByte,
       networkId = 1,
       forkBlockNumbers = ForkBlockNumbers(
         frontierBlockNumber = 0,
